feat: 변경된 로직으로 인터셉터, 스케줄러, 컨트롤러 적용 - 인터셉터: 인가방식 redis 조회로 변경 - 스케줄러   - 통과: 변경된 로직으로 대체   - 만료: 삭제(Redis TTL 로 대체) - 컨트롤러: 변경된 로직으로 대체

+++ src/main/java/com/chaw/concert/app/domain/concert/queue/usecase/PassWaitTokenUseCase.java
@@ -23,6 +23,9 @@ public class PassWaitTokenUseCase {
 
     public Output execute() {
         List<WaitToken> tokens = waitTokenRepository.getTokensEligibleForPass(PASS_SIZE);
+        if (tokens.isEmpty()) {
+            return new Output(0);
+        }
         waitTokenRepository.removeTokensBeforeTimeStamp(tokens.get(tokens.size() - 1).score());
         tokens.forEach(t -> {
             Long userId = Long.parseLong(t.key());
